+++ src/_types/etsy_payload_T.d.ts
@@ -1,6 +1,6 @@
 export interface etsy_payload_T<Params, Result> {
-	count:number,
-	results:Result[],
-	params:Params,
+	count:number
+	results:Result[]
+	params:Params
 	type:string
 }