+++ src/etsy_cache__put/index.d.ts
@@ -1,0 +1,4 @@
+export declare function etsy_cache__put():Promise<void>
+export {
+	etsy_cache__put as put_etsy_cache,
+}