+++ src/loading__s3__etsy_listing_a__/index.d.ts
@@ -1,0 +1,11 @@
+import type { WritableAtom_ } from '@ctx-core/nanostores'
+import type { Be, Ctx } from '@ctx-core/object'
+import type { etsy_listing_T } from '../_types'
+export declare const loading__s3__etsy_listing_a__:Be<
+	WritableAtom_<etsy_listing_T[]|undefined>
+>
+export {
+	loading__s3__etsy_listing_a__ as loading_s3_etsy_listing_a__,
+	loading__s3__etsy_listing_a__ as loading__s3_etsy_listing_a__,
+}
+export default function loading__s3__etsy_listing_a__reload(ctx:Ctx):Promise<void>