+++ src/loading__s3__etsy_listing_image_a__/index.d.ts
@@ -1,0 +1,11 @@
+import type { WritableAtom_ } from '@ctx-core/nanostores'
+import type { B, Ctx } from '@ctx-core/object'
+import type { etsy_listing_image_T } from '../_types'
+export declare const loading__s3__etsy_listing_image_a__:B<
+	WritableAtom_<etsy_listing_image_T[]|undefined>
+>
+export {
+	loading__s3__etsy_listing_image_a__ as loading_s3_etsy_listing_image_a__,
+	loading__s3__etsy_listing_image_a__ as loading_s3_etsy_listing_image_a$_,
+}
+export default function loading__s3__etsy_listing_image_a__reload(ctx:Ctx):Promise<void>