+++ src/s3_etsy_listing_a__.d.ts
@@ -1,5 +1,0 @@
-import { type ReadableAtom_ } from '@ctx-core/nanostores'
-import { type B } from '@ctx-core/object'
-import type { etsy_listing_T } from './_types'
-export declare const s3_etsy_listing_a__:B<ReadableAtom_<etsy_listing_T[]|undefined>>
-export { s3_etsy_listing_a__ as s3_etsy_listing_a$_ }